# backend/openapi_generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGRO",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/batch/": {
      "post": {
        "tags": [
          "Batches"
        ],
        "summary": "Create",
        "description": "Create a new batch entry.\n\nArgs:\n    entry (BatchCreate): Batch creation data.\n    db (Session): Database session dependency.\n    current_user (User): Current authenticated user.\n\nReturns:\n    BatchRead: The created batch object.",
        "operationId": "create_v1_batch__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Batches"
        ],
        "summary": "List batches",
        "description": "Get all batch entries.\n\nArgs:\n    skip (int): Number of records to skip.\n    limit (int): Maximum number of records to return.\n    db (Session): Database session dependency.\n\nReturns:\n    List[BatchRead]: List of batch objects.",
        "operationId": "read_all_v1_batch__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchRead"
                  },
                  "title": "Response Read All V1 Batch  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batch/by-item/{item_id}": {
      "get": {
        "tags": [
          "Batches"
        ],
        "summary": "List batches by item",
        "description": "Get batches for a specific item with available quantity.\n\nArgs:\n    item_id (int): The ID of the item.\n    db (Session): Database session dependency.\n\nReturns:\n    List[BatchRead]: List of batch objects for the item.",
        "operationId": "get_batches_by_item_v1_batch_by_item__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchRead"
                  },
                  "title": "Response Get Batches By Item V1 Batch By Item  Item Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batch/{batch_id}": {
      "get": {
        "tags": [
          "Batches"
        ],
        "summary": "Get batch by ID",
        "description": "Get a batch by ID.\n\nArgs:\n    batch_id (int): The ID of the batch to retrieve.\n    db (Session): Database session dependency.\n\nReturns:\n    BatchRead: The batch object.\n\nRaises:\n    AppException: If the batch is not found (404).",
        "operationId": "read_one_v1_batch__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Batches"
        ],
        "summary": "Update batch",
        "description": "Update a batch by ID.\n\nArgs:\n    batch_id (int): The ID of the batch to update.\n    entry_update (BatchUpdate): Fields to update.\n    db (Session): Database session dependency.\n    current_user (User): Current authenticated user.\n\nReturns:\n    BatchRead: The updated batch object.\n\nRaises:\n    AppException: If the batch is not found (404).",
        "operationId": "update_v1_batch__batch_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Batches"
        ],
        "summary": "Delete batch",
        "description": "Delete a batch.\n\nArgs:\n    batch_id (int): The ID of the batch to delete.\n    db (Session): Database session dependency.\n\nReturns:\n    None\n\nRaises:\n    AppException: If the batch is not found (404).",
        "operationId": "delete_v1_batch__batch_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/batch/batch/": {
      "post": {
        "tags": [
          "Batch",
          "Batches"
        ],
        "summary": "Create",
        "description": "Create a new batch entry.\n\nArgs:\n    entry (BatchCreate): Batch creation data.\n    db (Session): Database session dependency.\n    current_user (User): Current authenticated user.\n\nReturns:\n    BatchRead: The created batch object.",
        "operationId": "create_v1_batch_batch__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Batch",
          "Batches"
        ],
        "summary": "List batches",
        "description": "Get all batch entries.\n\nArgs:\n    skip (int): Number of records to skip.\n    limit (int): Maximum number of records to return.\n    db (Session): Database session dependency.\n\nReturns:\n    List[BatchRead]: List of batch objects.",
        "operationId": "read_all_v1_batch_batch__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchRead"
                  },
                  "title": "Response Read All V1 Batch Batch  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batch/batch/by-item/{item_id}": {
      "get": {
        "tags": [
          "Batch",
          "Batches"
        ],
        "summary": "List batches by item",
        "description": "Get batches for a specific item with available quantity.\n\nArgs:\n    item_id (int): The ID of the item.\n    db (Session): Database session dependency.\n\nReturns:\n    List[BatchRead]: List of batch objects for the item.",
        "operationId": "get_batches_by_item_v1_batch_batch_by_item__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchRead"
                  },
                  "title": "Response Get Batches By Item V1 Batch Batch By Item  Item Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batch/batch/{batch_id}": {
      "get": {
        "tags": [
          "Batch",
          "Batches"
        ],
        "summary": "Get batch by ID",
        "description": "Get a batch by ID.\n\nArgs:\n    batch_id (int): The ID of the batch to retrieve.\n    db (Session): Database session dependency.\n\nReturns:\n    BatchRead: The batch object.\n\nRaises:\n    AppException: If the batch is not found (404).",
        "operationId": "read_one_v1_batch_batch__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Batch",
          "Batches"
        ],
        "summary": "Update batch",
        "description": "Update a batch by ID.\n\nArgs:\n    batch_id (int): The ID of the batch to update.\n    entry_update (BatchUpdate): Fields to update.\n    db (Session): Database session dependency.\n    current_user (User): Current authenticated user.\n\nReturns:\n    BatchRead: The updated batch object.\n\nRaises:\n    AppException: If the batch is not found (404).",
        "operationId": "update_v1_batch_batch__batch_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Batch",
          "Batches"
        ],
        "summary": "Delete batch",
        "description": "Delete a batch.\n\nArgs:\n    batch_id (int): The ID of the batch to delete.\n    db (Session): Database session dependency.\n\nReturns:\n    None\n\nRaises:\n    AppException: If the batch is not found (404).",
        "operationId": "delete_v1_batch_batch__batch_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchCreate": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "quantity",
          "unit"
        ],
        "title": "BatchCreate"
      },
...
      "BatchRead": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "item_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "quantity",
          "unit",
          "id"
        ],
        "title": "BatchRead"
      },
      "BatchUpdate": {
        "properties": {
          "quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "title": "BatchUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}